Get Information about Basic Object

These functions are used to access the underlying properties of a Basic object.

get_type(x)

get_args(x)

get_hash(x)

get_str(x)

free_symbols(x)

function_symbols(x)

get_name(x)

get_prec(x)

Arguments

x

A Basic object.

Value

get_type(), get_hash(), get_str(), get_name() return a string.

get_args(), free_symbols(), function_symbols() return a VecBasic S4 object.

get_prec() returns an integer.

Details

get_type

Return the internal type

get_args

Return the internal arguments of a Basic object as a VecBasic

get_hash

Return the hash as a string

get_str

Return the string representation of the Basic object

free_symbols

Return free symbols in an expression

function_symbols

Return function symbols in an expression

get_name

Return name of a Basic object of type FunctionSymbol

get_prec

Return precision of a Basic object of type RealMPFR